pyproject.toml
[project]
name = "pcell"
version = "0.1.0"
description = "Package for analyzing place cells"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "xarray>=2025.10.1",
    "zarr>=2.17.0",
    "scipy>=1.10.0",
    "click>=8.0.0",
    "plotly>=5.0.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.0",
    "oasis-deconv>=0.1.0",
    "tqdm>=4.65.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build]
packages = ["pcell"]

[project.scripts]
pcell = "pcell.cli:pcell"

